requirements.txt
annotated-types==0.7.0
anyio==4.4.0
blinker==1.8.2
certifi==2024.6.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
ollama==0.4.7
packaging==24.1
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
python-telegram-bot==21.3
requests==2.31.0
sniffio==1.3.1
typing_extensions==4.12.2
Werkzeug==3.0.3
